Move FingerprintStorage,FileFingerprintStorage -> jenkins.fingerprints package
